pyproject.toml
[project]
name = "gauntlet-graphs-assignment"
version = "0.1.0"
description = "Graph-based automation for premises data processing with Google Places API and duplicate detection"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langgraph>=0.2",
    "pymysql>=1.1",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "googlemaps>=4.10",
    "pydantic>=2.10",
    "jupyter>=1.0",
    "pytest>=8.3",
]
